fix: move Console to utils/tools, so cli/utils isnt build into main entry
